* gh-43941:
  Change return type for VirtualThreadsInfo.getMounted() to int

Closes gh-43941

+++ spring-boot-project/spring-boot/src/main/java/org/springframework/boot/info/ProcessInfo.java
@@ -151,7 +151,7 @@ public class ProcessInfo {
 			this.poolSize = poolSize;
 		}
 
-		public long getMounted() {
+		public int getMounted() {
 			return this.mounted;
 		}
